requirements.txt
streamlit==1.32.0
langchain==0.1.12
python-dotenv==1.0.1
requests==2.31.0
langchain-groq==0.0.1
python-dateutil==2.8.2
pydantic>=2.0.0
aiohttp==3.9.3 